+++ src/hooks/useScrolled.ts
@@ -5,7 +5,7 @@ export function useScrolled(elementId: string) {
   const [isScrolled, setIsScrolled] = useState(true);
 
   useEffect(() => {
-    let DELAY_STICKY_TRANSITION = window.innerWidth < 600 ? 700 : 300;
+    let DELAY_STICKY_TRANSITION = window.innerWidth < 767 ? 0 : 300;
     function handleScroll() {
       const element = document.getElementById(elementId);
       if (element) {
@@ -18,7 +18,7 @@ export function useScrolled(elementId: string) {
       }
     }
     function handleScreenSize() {
-      const newDelay = window.innerWidth < 600 ? 700 : 300;
+      const newDelay = window.innerWidth < 767 ? 0 : 300;
       if (newDelay !== DELAY_STICKY_TRANSITION) {
         setIsScrolled(true);
         DELAY_STICKY_TRANSITION = newDelay;